Exclude vendor directories from rubocop
Apparently with the bundler cache that ruby/setup-ruby provides
it is common to pull in vendor'd rubocop.yml files even if you
did not intend to.

This change informs rubocop to ignore these files, even though this
normally means "do not consider this file for scanning for style fixes".

For more info : https://github.com/rubocop/rubocop/issues/9832