Disable CI on pull requests
This would trigger deploys to heroku and yeah I'm not doing that.
Doesn't immediately seem possible to tell GH actions to avoid
running a job if it's from a pull request so we can take this
out until such a time that we actually need to support PRs.

HashTag Do Things That Don't Scale