Remove flakiness from tests
The test for friday was passing 50% of the time on CI, and
I noticed that it would pass iff Alice was the MC for both weeks.
It's hard to make this completely deterministic, because of the fact
that randomness is introduced by the backend.

Adding more backmakers to the team reduces the chance of this happening.

Apparently the cause for the false positive was if Alice was already
the MC, and we submit, it never actually changes anything, because
there was a javascript error on the page. It seems in the test
environment that there is no CSRF token. Strange.