Fix turbolinks + CSRF bug
Turbolinks greatly speeds up the page loading time, but it's not magic.
In particular when a page includes a form and uses a CSRF authenticity token
anytime the user submits that form after having navigated using turbolinks
the form submission will fail, because the CSRF token in the <head> will no longer
match the value in the <form>.

A few possible solutions :

* disable turbolinks in production (eh maybe ?)
* disable CSRF (no way, josé)
* use JS to dynamically patch the <input name=authenticity_token> (wild shenanigans)
* disable turbolinks for that specific page (least bad, but also least harmonious)

I'm opting for the fourth option because it's the easiest to put in place for now.